+++ rest/internal/wrap/wrap.go
@@ -22,9 +22,6 @@ func disabledErr[R model.Release]() error {
 }
 
 func Generic[R model.Release](api any) (capabilities.GenericCapabilities, error) {
-	if g, ok := api.(capabilities.GenericCapabilities); ok {
-		return g, nil
-	}
 	var r R
 	switch any(r).(type) {
 	case model.R4: